refactor(predict-next): organize canonical read contracts
