+++ app/components/UI/PredictNext/queries/marketDataQueries.test.ts
