+++ app/components/UI/PredictNext/queries/marketDataQueries.ts
